refactor(operations): :recycle: Splitted dialog in components
+ loading in state
