+++ store/operations/getters.ts
@@ -10,9 +10,16 @@ const getters: GetterTree<OperationState, RootState> = {
    * Get the operations for the auther user in the selected account
    *
    * @param state The state
-   * @returns The operation
+   * @returns The operations
    */
   getOperations: (state) => state,
+  /**
+   * Get the loading state
+   *
+   * @param state The state
+   * @returns The loading state
+   */
+  getLoadingState: (state) => state.loading,
 }
 
 export default getters
